+++ al-khaser/Anti VM/Generic.h
@@ -32,3 +32,4 @@ BOOL serial_number_bios_wmi();
 BOOL model_computer_system_wmi();
 BOOL manufacturer_computer_system_wmi();
 BOOL current_temperature_acpi_wmi();
+BOOL process_id_processor_wmi();